Refactoring for debugging tests (#682)
* Make test parameterized and add description
* Use side-by-side scenario
* Test with different api
* Switch order for api and breakpoint
* Move test_debug_dppy_numba.py to numba_dppy/tests/debugging
* Extract tests/debugging/gdb.py
* Move app() to conftests.py
* Move skip_no_numba055 to tests/_helper.py
* Move skip_no_gdb to tests/_helper.py
* Extract tests for breakpoints in tests/debugging/test_breakpoints.py
* Move test_break_conditional to test_breakpoints.py
* Move tests to test_breakpoints.py
- test_break_file_function
- test_break_function
- test_break_nested_function
Also improved docstrings for tests.
* Add docstring for test_breakpoints.py module
* Add docstring for test_break_conditional
* Add common_loop_body_native_function_name
* Add test_side_by_side.py
* Add test_stepping.py
* Fix black and pre-commit
* Add test_info.py
* Add test_backtraces.py
* Add file:function case to test_breakpoint_with_condition_by_function_argument
* Add breakpoint_api_cases
* Add test_breakpoint_common
* Replace test_break_function with parameter
* Replace test_break_nested_function with parameter
* Add test_breakpoint_with_condition_common
* Add common.py with script_path() and breakpoint_by_mark()
* Add test_common.py
* Add side_by_side_breakpoint
* Use marks for simple_sum.py
* Add breakpoint_by_function()
* Add setup_breakpoint()
* Use setup_breakpoint
* Remove expected_line
* Use setup_breakpoint
* Skip if no GDB for test_common.py
* Skip tests with app() if pexpect not installed
