+++ numba_dppy/examples/debug/side-by-side.py
@@ -22,7 +22,7 @@ import numba_dppy as dppy
 
 
 def common_loop_body(param_a, param_b):
-    param_c = param_a + 10  # Set breakpoint
+    param_c = param_a + 10  # Set breakpoint here
     param_d = param_b * 0.5
     result = param_c + param_d
     return result
